+++ liana/testing/__init__.py
@@ -1,6 +1,6 @@
 from ._sample_lrs import sample_lrs
 from ._sample_anndata import generate_toy_spatial
-from ._toy_adata import get_toy_adata
+from ._sample_anndata import generate_toy_adata
 from .datasets import kang_2018
 from ._sample_misty import _sample_interactions
 from ._sample_misty import _sample_target_metrics
