+++ liana/testing/_sample_anndata.py
@@ -1,8 +1,10 @@
 import numpy as np
+from numpy import random
+from scanpy.datasets import pbmc68k_reduced
 
 def generate_toy_spatial():
     from scanpy.datasets import pbmc68k_reduced
-    from liana.method.sp._spatial_neighbors import spatial_neighbors
+    from liana.utils._spatial_neighbors import spatial_neighbors
     
     adata = pbmc68k_reduced()
     
@@ -33,3 +35,18 @@ def generate_toy_mdata():
     mdata.obsm = adata.obsm
     
     return mdata
+
+
+def generate_toy_adata():
+    adata = pbmc68k_reduced()
+    sample_key = 'sample'
+
+    rng = random.default_rng(0)
+
+    # create fake samples
+    adata.obs[sample_key] = rng.choice(['A', 'B', 'C', 'D'], size=len(adata.obs))
+
+    # group samples into conditions
+    adata.obs['case'] = adata.obs[sample_key].map({'A': 'yes', 'B': 'yes', 'C': 'no', 'D': 'no'})
+
+    return adata
