+++ liana/utils/__init__.py
@@ -1,3 +1,4 @@
 from liana.utils.obsm_to_adata import obsm_to_adata
 from liana.utils.mdata_to_anndata import mdata_to_anndata
 from liana.utils.transform import zi_minmax, neg_to_zero
+from liana.utils._spatial_neighbors import spatial_neighbors
