+++ liana/utils/_spatial_neighbors.py
